chore|refactor [RTP] app config & refactoring
- Changes bundler to vite
- Generates client from openapi
- Removes parcel and related code
- Removes unused code and files
- Fixes typescript errors
- Refactors i18n typing
- Fixes imports
