+++ .env
@@ -1,3 +1,3 @@
 ENV="LOCAL"
 VERSION="test"
-API_URL="https://api-rtp.dev.cstar.pagopa.it/rtp/v1"
+VITE_API_URL="https://api-rtp.dev.cstar.pagopa.it/rtp/v1"
