+++ .gitignore
@@ -25,3 +25,5 @@ yarn-error.log*
 .parcel-cache
 dist
 build
+
+/generated
