--- .parcelrc
@@ -1,13 +1,0 @@
-{
-  "extends": [
-    "@parcel/config-default"
-  ],
-  "reporters": [
-    "...",
-    "parcel-reporter-static-files-copy"
-  ],
-  "resolvers": [
-    "parcel-resolver-ignore",
-    "..."
-  ]
-}
