+++ index.html
@@ -54,6 +54,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script src="./index.tsx" type="module"></script>
+    <script src="/src/main.tsx" type="module"></script>
   </body>
 </html>
