+++ package.json
@@ -1,79 +1,75 @@
 {
-  "name": "rtp-start-app",
-  "version": "0.1.0",
-  "private": true,
-  "source": "src/index.html",
-  "dependencies": {
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^6.1.5",
-    "@mui/lab": "^6.0.0-beta.13",
-    "@mui/material": "^6.1.5",
-    "@mui/x-date-pickers": "^7.21.0",
-    "@pagopa/mui-italia": "^1.5.1-0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.7.7",
-    "dayjs": "^1.11.13",
-    "formik": "^2.4.6",
-    "i18next": "^23.16.3",
-    "i18next-browser-languagedetector": "^8.0.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-i18next": "^15.1.0",
-    "react-router-dom": "^6.26.2",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "generate": "",
-    "start": "npm run generate && parcel --no-cache",
-    "build": "npm run generate && tsc & parcel build",
-    "test": "vitest",
-    "coverage": "vitest --coverage"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^6.6.1",
-    "@testing-library/react": "^16.0.1",
-    "@types/jest": "^29.5.14",
-    "@types/react": "^18.3.11",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^8.8.1",
-    "@typescript-eslint/parser": "^8.8.1",
-    "@vitejs/plugin-react": "^4.3.2",
-    "dotenv": "^16.4.5",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-promise": "^6.6.0",
-    "jsdom": "^25.0.1",
-    "parcel": "^2.12.0",
-    "parcel-reporter-static-files-copy": "^1.5.3",
-    "parcel-resolver-ignore": "^2.2.0",
-    "process": "^0.11.10",
-    "rimraf": "^6.0.1",
-    "typescript": "^5.6.2",
-    "vite": "^5.4.8",
-    "vite-tsconfig-paths": "^5.0.1",
-    "vitest": "^2.1.3"
-  }
+	"name": "rtp-start-app",
+	"version": "0.1.0",
+	"private": true,
+	"source": "src/index.html",
+	"dependencies": {
+		"@emotion/react": "^11.13.3",
+		"@emotion/styled": "^11.13.0",
+		"@mui/icons-material": "^6.1.5",
+		"@mui/lab": "^6.0.0-beta.13",
+		"@mui/material": "^6.1.5",
+		"@mui/x-date-pickers": "^7.21.0",
+		"@pagopa/mui-italia": "^1.5.1-0",
+		"@testing-library/user-event": "^13.5.0",
+		"axios": "^1.7.7",
+		"dayjs": "^1.11.13",
+		"formik": "^2.4.6",
+		"i18next": "^23.16.3",
+		"i18next-browser-languagedetector": "^8.0.0",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-i18next": "^15.1.0",
+		"react-router-dom": "^6.26.2",
+		"web-vitals": "^2.1.4"
+	},
+	"scripts": {
+		"generate-api-client": "swagger-typescript-api -p https://raw.githubusercontent.com/pagopa/rtp-apis/refs/heads/main/src/rtp/api/pagopa/send.openapi.yaml -o ./generated -n apiClient --axios",
+		"generate-data-contracts": "swagger-typescript-api  --modular --no-client -p https://raw.githubusercontent.com/pagopa/p4pa-pu-bff/refs/heads/develop/openapi/p4pa-pu-bff.openapi.yaml -o ./generated",
+		"generate": "yarn generate-api-client && yarn generate-data-contracts",
+		"start": "yarn generate && vite",
+		"build": "yarn generate && tsc -b && vite build",
+		"preview": "vite preview",
+		"test": "vitest"
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@testing-library/jest-dom": "^6.6.1",
+		"@testing-library/react": "^16.0.1",
+		"@types/jest": "^29.5.14",
+		"@types/react": "^18.3.11",
+		"@types/react-dom": "^18.3.0",
+		"@typescript-eslint/eslint-plugin": "^8.8.1",
+		"@typescript-eslint/parser": "^8.8.1",
+		"@vitejs/plugin-react": "^4.3.2",
+		"dotenv": "^16.4.5",
+		"eslint": "^8.57.1",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-config-standard": "^17.1.0",
+		"eslint-plugin-import": "^2.31.0",
+		"eslint-plugin-n": "^16.6.2",
+		"eslint-plugin-prettier": "^5.2.1",
+		"eslint-plugin-promise": "^6.6.0",
+		"jsdom": "^25.0.1",
+		"process": "^0.11.10",
+		"rimraf": "^6.0.1",
+		"swagger-typescript-api": "^13.0.23",
+		"typescript": "^5.6.2",
+		"vite": "^6.0.11",
+		"vite-tsconfig-paths": "^5.0.1",
+		"vitest": "^3.0.4",
+		"vitest-dom": "^0.1.1",
+		"zodock": "^0.1.0"
+	}
 }
