--- src/global.d.ts
@@ -1,2 +1,0 @@
-declare module '*.png';
-declare module '*.svg';
