+++ src/vite-env.d.ts
@@ -1,0 +1,8 @@
+/// <reference types="vite/client" />
+interface ImportMetaEnv {
+    readonly VITE_DEPLOY_PATH: string
+  }
+  
+  interface ImportMeta {
+    readonly env: ImportMetaEnv
+  }
