+++ tsconfig.json
@@ -1,42 +1,36 @@
 {
-    "compilerOptions": {
-      "allowJs": false,
-      "allowSyntheticDefaultImports": true,
-      "allowUmdGlobalAccess": true,
-      "baseUrl": "src",
-      "esModuleInterop": true,
-      "experimentalDecorators": true,
-      "forceConsistentCasingInFileNames": true,
-      "importHelpers": true,
-      "jsx": "react",
-      "lib": ["es6", "dom", "esnext.asynciterable", "es5"],
-      "module": "esnext",
-      "moduleResolution": "node",
-      "noEmit": true,
-      "noImplicitAny": true,
-      "noImplicitReturns": true,
-      "noUnusedLocals": true,
-      "noUnusedParameters": true,
-      "preserveConstEnums": true,
-      "removeComments": true,
-      "resolveJsonModule": true,
-      "skipLibCheck": true,
-      "strictFunctionTypes": true,
-      "strictNullChecks": true,
-      "strict": true,
-      "target": "ES2020",
-      //"types": ["vitest/globals"]
-    },
-    "include": [
-      "src"
-    ],
-    "exclude": [
-      "node_modules",
-      "build",
-      "scripts",
-      "acceptance-tests",
-      "webpack",
-      "vitest",
-      "src/setupTests.ts"
-    ]
+	"compilerOptions": {
+		"baseUrl": ".",
+		"paths": {
+			"src/*": [
+				"./src/*"
+			]
+		},
+		"tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
+		"target": "ES2020",
+		"useDefineForClassFields": true,
+		"lib": [
+			"ES2020",
+			"DOM",
+			"DOM.Iterable"
+		],
+		"module": "ESNext",
+		"skipLibCheck": true,
+		/* Bundler mode */
+		"moduleResolution": "Bundler",
+		"allowImportingTsExtensions": true,
+		"isolatedModules": true,
+		"moduleDetection": "force",
+		"noEmit": true,
+		"jsx": "react-jsx",
+		/* Linting */
+		"strict": true,
+		"noUnusedLocals": true,
+		"noUnusedParameters": true,
+		"noFallthroughCasesInSwitch": true,
+		"noUncheckedSideEffectImports": true
+	},
+	"include": [
+		"src"
+	]
 }
